split the functions used by languages and file extentions
