+++ tests/cli/output/replace.txt
@@ -1,2 +1,2 @@
-do_source_file(1526): Parsing: input/I-3310.c as language C
-do_source_file(1526): Parsing: input/backup.h as language C-Header
+do_source_file(1521): Parsing: input/I-3310.c as language C
+do_source_file(1521): Parsing: input/backup.h as language C-Header
